refactor : Operator Enum 분리
